+++ lib/hosting.ts
@@ -3,7 +3,7 @@ import { Construct } from "constructs";
 import { Bucket, type IBucket, BlockPublicAccess, ObjectOwnership, BucketEncryption } from "aws-cdk-lib/aws-s3";
 import { Role, PolicyStatement, Effect, ServicePrincipal, AnyPrincipal, ManagedPolicy } from "aws-cdk-lib/aws-iam";
 import { S3Origin } from "aws-cdk-lib/aws-cloudfront-origins";
-import { Distribution, type IDistribution, CachePolicy, SecurityPolicyProtocol, HttpVersion, ResponseHeadersPolicy, HeadersFrameOption, HeadersReferrerPolicy, type BehaviorOptions, AllowedMethods, ViewerProtocolPolicy, CacheCookieBehavior, CacheHeaderBehavior, CacheQueryStringBehavior, CfnOriginAccessControl, Function as CloudFrontFunction, FunctionCode as CloudFrontFunctionCode, FunctionEventType, OriginAccessIdentity, CachedMethods, FunctionRuntime, LambdaEdgeEventType, experimental } from "aws-cdk-lib/aws-cloudfront";
+import { Distribution, CachePolicy, SecurityPolicyProtocol, HttpVersion, ResponseHeadersPolicy, HeadersFrameOption, HeadersReferrerPolicy, type BehaviorOptions, AllowedMethods, ViewerProtocolPolicy, CacheCookieBehavior, CacheHeaderBehavior, CacheQueryStringBehavior, CfnOriginAccessControl, CachedMethods, LambdaEdgeEventType, experimental, OriginRequestPolicy, OriginRequestHeaderBehavior, OriginRequestCookieBehavior, OriginRequestQueryStringBehavior, OriginProtocolPolicy } from "aws-cdk-lib/aws-cloudfront";
 import { AaaaRecord, ARecord, HostedZone, type IHostedZone, RecordTarget } from "aws-cdk-lib/aws-route53";
 import { CloudFrontTarget } from "aws-cdk-lib/aws-route53-targets";
 import { Certificate } from "aws-cdk-lib/aws-certificatemanager";
@@ -15,6 +15,7 @@ export interface HostingProps {
     domain?: string;
     globalCertificateArn?: string;
     hostedZoneId?: string;
+    errorPagePath?: string;
     redirects?: { source: string; destination: string; }[];
     rewrites?: { source: string; destination: string; }[];
     headers?: { path: string; name: string; value: string; }[];
@@ -39,7 +40,7 @@ export class HostingConstruct extends Construct {
     /**
      * The CloudFront distribution.
      */
-    public distribution:  IDistribution;
+    public distribution:  Distribution;
 
     /**
      * The CloudFront distribution origin that routes to S3 HTTP server.
@@ -316,11 +317,10 @@ export class HostingConstruct extends Construct {
      * @private
      */
     private createHostingBucket(props: HostingProps) {
-        let originLogsBucket: Bucket | undefined;
 
         // Hosting bucket access log bucket
-        if (props.debug) {
-          originLogsBucket = new Bucket(this, "OriginLogsBucket", {
+        const originLogsBucket = props.debug
+          ? new Bucket(this, "OriginLogsBucket", {
             bucketName: `${props.resourceIdPrefix}-origin-logs`,
             encryption: BucketEncryption.S3_MANAGED,
             blockPublicAccess: BlockPublicAccess.BLOCK_ALL,
@@ -328,8 +328,8 @@ export class HostingConstruct extends Construct {
             enforceSSL: true,
             removalPolicy: RemovalPolicy.DESTROY,
             autoDeleteObjects: true
-          });
-        }
+          })
+          : undefined;
 
         // primary hosting bucket
         const bucket = new Bucket(this, "HostingBucket", {
@@ -388,8 +388,8 @@ export class HostingConstruct extends Construct {
     private createCloudfrontDistribution(props: HostingProps) {
 
         // access logs bucket
-        if (props.debug) {
-          this.accessLogsBucket = new Bucket(this, "AccessLogsBucket", {
+        this.accessLogsBucket = props.debug
+          ? new Bucket(this, "AccessLogsBucket", {
             bucketName: `${props.resourceIdPrefix}-access-logs`,
             encryption: BucketEncryption.S3_MANAGED,
             blockPublicAccess: BlockPublicAccess.BLOCK_ALL,
@@ -397,32 +397,13 @@ export class HostingConstruct extends Construct {
             enforceSSL: true,
             removalPolicy: RemovalPolicy.DESTROY,
             autoDeleteObjects: true
-          });
-        }
+          })
+          : undefined;
 
-        // defaultCachePolicy
-        const defaultCachePolicy = new CachePolicy(this, "DefaultCachePolicy", {
-          cachePolicyName: `${props.resourceIdPrefix}-DefaultCachePolicy`,
-          comment: 'Cache policy for HTML documents with short TTL',
-          defaultTtl: Duration.minutes(1),
-          minTtl: Duration.seconds(0),
-          maxTtl: Duration.minutes(1),
-          headerBehavior: props.allowHeaders?.length
-            ? CacheHeaderBehavior.allowList(...props.allowHeaders)
-            : CacheHeaderBehavior.none(),
-          cookieBehavior: props.allowCookies?.length
-            ? CacheCookieBehavior.allowList(...props.allowCookies)
-            : CacheCookieBehavior.none(),
-          queryStringBehavior: props.allowQueryParams?.length 
-            ? CacheQueryStringBehavior.allowList(...props.allowQueryParams) 
-            : (props.denyQueryParams?.length 
-              ? CacheQueryStringBehavior.denyList(...props.denyQueryParams) 
-              : CacheQueryStringBehavior.none()),
-          enableAcceptEncodingGzip: true,
-          enableAcceptEncodingBrotli: true,
-        });
-
-        // ResponseHeadersPolicy
+        /**
+         * Response Headers Policy
+         * This policy is used to set default security headers for the CloudFront distribution.
+         */
         const responseHeadersPolicy = new ResponseHeadersPolicy(this, "ResponseHeadersPolicy", {
           responseHeadersPolicyName: `${props.resourceIdPrefix}-ResponseHeadersPolicy`,
           comment: "ResponseHeadersPolicy" + Aws.STACK_NAME + "-" + Aws.REGION,
@@ -468,8 +449,36 @@ export class HostingConstruct extends Construct {
           },
           removeHeaders: ['server', 'age' , 'date'],
         });
+
+        /**
+         * The default cache policy for HTML documents with short TTL.
+         * This policy is used for the default behavior of the CloudFront distribution.
+         */
+        const defaultCachePolicy = new CachePolicy(this, "DefaultCachePolicy", {
+          cachePolicyName: `${props.resourceIdPrefix}-DefaultCachePolicy`,
+          comment: 'Cache policy for HTML documents with short TTL',
+          defaultTtl: Duration.minutes(1),
+          minTtl: Duration.seconds(0),
+          maxTtl: Duration.minutes(1),
+          headerBehavior: props.allowHeaders?.length
+            ? CacheHeaderBehavior.allowList(...props.allowHeaders)
+            : CacheHeaderBehavior.none(),
+          cookieBehavior: props.allowCookies?.length
+            ? CacheCookieBehavior.allowList(...props.allowCookies)
+            : CacheCookieBehavior.none(),
+          queryStringBehavior: props.allowQueryParams?.length 
+            ? CacheQueryStringBehavior.allowList(...props.allowQueryParams) 
+            : (props.denyQueryParams?.length 
+              ? CacheQueryStringBehavior.denyList(...props.denyQueryParams) 
+              : CacheQueryStringBehavior.none()),
+          enableAcceptEncodingGzip: true,
+          enableAcceptEncodingBrotli: true,
+        });
       
-        // defaultBehavior
+        /**
+         * The default behavior for the CloudFront distribution.
+         * This behavior is used for the default behavior of the CloudFront distribution.
+         */
         const defaultBehavior: BehaviorOptions = {
           origin: this.s3Origin,
           compress: true,
@@ -490,18 +499,44 @@ export class HostingConstruct extends Construct {
           ],
         };
 
-        // staticAssetsBehaviour
+        // Additional behaviors
+        const additionalBehaviors: { [pathPattern: string]: BehaviorOptions } = {};
+
+        /**
+         * The behavior for static assets.
+         * It is configured to cache static assets for a longer period of time.
+         * Using a managed cache policy CACHING_OPTIMIZED.
+         * Using a managed response headers policy: CORS_ALLOW_ALL_ORIGINS_WITH_PREFLIGHT
+         */
         const staticAssetsBehaviour: BehaviorOptions = {
           origin: this.s3Origin,
           compress: true,
-          responseHeadersPolicy: responseHeadersPolicy,
+          responseHeadersPolicy: ResponseHeadersPolicy.CORS_ALLOW_ALL_ORIGINS_WITH_PREFLIGHT,
           cachePolicy: CachePolicy.CACHING_OPTIMIZED,
           allowedMethods: AllowedMethods.ALLOW_GET_HEAD_OPTIONS,
           cachedMethods: CachedMethods.CACHE_GET_HEAD_OPTIONS,
           viewerProtocolPolicy: ViewerProtocolPolicy.REDIRECT_TO_HTTPS,
         };
 
-        // finally, create distribution
+        // Add static asset behaviors
+        const staticAssetPatterns = [
+          '*.png',
+          '*.jpg',
+          '*.jpeg',
+          '*.gif',
+          '*.ico',
+          '*.css',
+          '*.js',
+        ];
+        
+        for (const pattern of staticAssetPatterns) {
+          additionalBehaviors[pattern] = staticAssetsBehaviour;
+        }
+    
+        /**
+         * Create CloudFront Distribution
+         * 
+         */
         const distributionName = `${props.resourceIdPrefix}-cdn`;
 
         const distributionProps = {
@@ -509,33 +544,17 @@ export class HostingConstruct extends Construct {
           enableLogging: props.debug ? true : false,
           logBucket: props.debug ? this.accessLogsBucket : undefined,
           defaultBehavior: defaultBehavior,
-          additionalBehaviors: {
-            "*.jpg": staticAssetsBehaviour,
-            "*.jpeg": staticAssetsBehaviour,
-            "*.png": staticAssetsBehaviour,
-            "*.gif": staticAssetsBehaviour,
-            "*.bmp": staticAssetsBehaviour,
-            "*.tiff": staticAssetsBehaviour,
-            "*.ico": staticAssetsBehaviour,
-            "*.js": staticAssetsBehaviour,
-            "*.css": staticAssetsBehaviour
-          },
+          additionalBehaviors: additionalBehaviors,
           responseHeadersPolicy: responseHeadersPolicy,
           httpVersion: HttpVersion.HTTP3,
           minimumProtocolVersion: SecurityPolicyProtocol.TLS_V1_2_2021,
           defaultRootObject: "index.html",
           errorResponses: [
             {
-              httpStatus: 403,
-              responseHttpStatus: 200,
-              ttl: Duration.seconds(0),
-              responsePagePath: '/index.html',
-            },
-            {
               httpStatus: 404,
               responseHttpStatus: 404,
               ttl: Duration.seconds(0),
-              responsePagePath: '/index.html',
+              responsePagePath: props.errorPagePath || '/index.html',
             },
           ],
           ...(props.domain && props.globalCertificateArn